Simplify ignore interval function, remove unneeded subtract interval tests, and add ignore interval test case. Add documentation on project structure. Refactor post-processing to separate module. Rewrite merge function to not heap allocate at end and not use vecdeque. Use better overlap_length function. Bump version.
