+++ .gitattributes
@@ -72,3 +72,4 @@ core/test/collapse/input/aln_1.bam filter=lfs diff=lfs merge=lfs -text
 core/test/inversion/input/aln_1.fa.gz filter=lfs diff=lfs merge=lfs -text
 core/test/inversion/input/aln_1.bam filter=lfs diff=lfs merge=lfs -text
 core/test/no_indel_merging/input/aln_1.fa.gz filter=lfs diff=lfs merge=lfs -text
+core/test/standard/expected/aln_1_ignore.bed filter=lfs diff=lfs merge=lfs -text
