+++ core/test/standard/expected/aln_1_ignore.bed
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:6ffce89909cf2de03a680a13155d3c0efb8c6d04cc15fc1d6fb237bf2fbe8488
+size 5100
