+++ core/test/standard/input/aln_1_ignore.bed
@@ -1,0 +1,2 @@
+haplotype2-0000133	7028106	7029969
+haplotype2-0000133	7029969	7146183
