+++ py/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "py-nucflag"
-version = "0.1.9"
+version = "0.1.10"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
